Extract mechanics review helpers
